refactoring plotting module name
