--- sklarpy/_plotting/__init__.py
@@ -1,2 +1,0 @@
-from sklarpy._plotting._pair_plot import pair_plot
-from sklarpy._plotting._threeD_plot import threeD_plot
